+++ server/api/moments/pending/[group_id]/[image_id]/index.patch.ts
@@ -24,7 +24,9 @@ export default defineSupabaseEventHandler(async (event, user, client, server) =>
 	************************************************************************************
 	*/
 
-	
+	const { data: postData, error: postError } = await server.from("posts").select("*").eq("id", image_id).select().single<Tables<"posts">>()
+	if (postError) return useReturnResponse(event, internalServerError);
+
 	if (!request.has_been_accepted) {
 		await client.from('posts').delete().eq('id', image_id)
 
@@ -43,6 +45,19 @@ export default defineSupabaseEventHandler(async (event, user, client, server) =>
 			last_action: "Approved"
 		}).eq("id", group_id)
 	}
+
+	const { error: logError } = await server.from("logbook").insert({
+		message: request.has_been_accepted ? 'Approved :member: photo' : 'Rejected :member: photo',
+		performed_by_id: user.id,
+		target_user_id: postData.author_id,
+		action_type: request.has_been_accepted ? "created" : "deleted",
+		group_id: group_id,
+		context: {
+			type: "image",
+		}
+	})
+
+	if (logError) return useReturnResponse(event, internalServerError)
 
 	/*
 	************************************************************************************
