+++ server/api/moments/pending/marked/[group_id]/index.patch.ts
@@ -25,7 +25,10 @@ export default defineSupabaseEventHandler(async (event, user, client, server) =>
 	*/
 
 	for (const [key, value] of Object.entries(request)) {
-		
+
+		const { data: postData, error: postError } = await server.from("posts").select("*").eq("id", key).select().single<Tables<"posts">>()
+		if (postError) return useReturnResponse(event, internalServerError);
+
 		if (!value) {
 			await client.from('posts').delete().eq('id', key)
 
@@ -44,6 +47,21 @@ export default defineSupabaseEventHandler(async (event, user, client, server) =>
 				last_action: "Approved"
 			}).eq("id", group_id)
 		}
+
+		const { error: logError } = await server.from("logbook").insert({
+			message: value 
+				? (postData.author_id === user.id ? 'Approved their own photo' : 'Approved :member: photo') 
+				: (postData.author_id === user.id ? 'Rejected their own photo' : 'Rejected :member: photo'),
+			performed_by_id: user.id,
+			target_user_id: postData.author_id,
+			action_type: value ? "created" : "deleted",
+			group_id: group_id,
+			context: {
+				type: "image",
+			}
+		})
+
+		if (logError) return useReturnResponse(event, internalServerError)
 	}
 
 	/*
